fix: useTimeWithDisabled
